Refactor to conform to the existing API design
