+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from './mime';
 export * from './ext';
-export * from './filekind';
 export * from './data';
